third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/defer-script/defer-script-xml-expected.txt

This is a testharness.js-based test.
[FAIL] Defer Script Execution Order
  Uncaught Error: assert_array_equals: Execution order expected property 2 to be "Inline2" but got "Defer1" (expected array ["Inline1", "Sync1", "Inline2", "Sync2", "EndOfBody", "Defer1", "Defer2", "DOMContentLoaded"] got ["Inline1", "Sync1", "Defer1", "Inline2", "Defer2", "Sync2", "EndOfBody", "DOMContentLoaded"])
Harness: the test ran to completion.
